+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/io/autoware_osm_parser.hpp
