+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/regulatory_elements/autoware_traffic_light.hpp
