+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/regulatory_elements/no_stopping_area.hpp
