+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/regulatory_elements/road_marking.hpp
