+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/regulatory_elements/virtual_traffic_light.hpp
