+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/utility/message_conversion.hpp
