+++ autoware_lanelet2_extension/include/autoware_lanelet2_extension/utility/utilities.hpp
