+++ autoware_lanelet2_extension_python/autoware_lanelet2_extension_python/projection/__init__.py
@@ -1,0 +1,3 @@
+import autoware_lanelet2_extension_python._autoware_lanelet2_extension_python_boost_python_projection as _projection_cpp
+
+MGRSProjector = _projection_cpp.MGRSProjector
