+++ autoware_lanelet2_extension_python/autoware_lanelet2_extension_python/regulatory_elements/__init__.py
@@ -1,4 +1,4 @@
-import lanelet2_extension_python._lanelet2_extension_python_boost_python_regulatory_elements as _regulatory_elements_cpp
+import autoware_lanelet2_extension_python._autoware_lanelet2_extension_python_boost_python_regulatory_elements as _regulatory_elements_cpp
 
 AutowareTrafficLight = _regulatory_elements_cpp.AutowareTrafficLight
 Crosswalk = _regulatory_elements_cpp.Crosswalk
