+++ autoware_lanelet2_extension_python/autoware_lanelet2_extension_python/utility/__init__.py
@@ -1,0 +1,1 @@
+import autoware_lanelet2_extension_python.regulatory_elements  # noqa
