+++ autoware_lanelet2_extension_python/autoware_lanelet2_extension_python/utility/query.py
@@ -1,7 +1,7 @@
 from geometry_msgs.msg import Point
 from geometry_msgs.msg import Pose
 import lanelet2.core
-import lanelet2_extension_python._lanelet2_extension_python_boost_python_utility as _utility_cpp
+import autoware_lanelet2_extension_python._autoware_lanelet2_extension_python_boost_python_utility as _utility_cpp
 
 # from rclpy.serialization import deserialize_message
 from rclpy.serialization import serialize_message
