+++ autoware_lanelet2_extension_python/autoware_lanelet2_extension_python/utility/utilities.py
@@ -2,7 +2,7 @@ from geometry_msgs.msg import Point
 from geometry_msgs.msg import Pose
 from geometry_msgs.msg import Quaternion
 import lanelet2
-import lanelet2_extension_python._lanelet2_extension_python_boost_python_utility as _utility_cpp
+import autoware_lanelet2_extension_python._autoware_lanelet2_extension_python_boost_python_utility as _utility_cpp
 from rclpy.serialization import serialize_message
 
 combineLaneletsShape = _utility_cpp.combineLaneletsShape
