--- autoware_lanelet2_extension_python/autowarelanelet2_extension_python/projection/__init__.py
@@ -1,3 +1,0 @@
-import lanelet2_extension_python._lanelet2_extension_python_boost_python_projection as _projection_cpp
-
-MGRSProjector = _projection_cpp.MGRSProjector
