--- autoware_lanelet2_extension_python/autowarelanelet2_extension_python/utility/__init__.py
@@ -1,1 +1,0 @@
-import lanelet2_extension_python.regulatory_elements  # noqa
